pyproject.toml
[project]
name = "ttsutil"
version = "0.2.0"
description = "Utilities for creating, updating, and managing sets of Text-To-Speech files"
authors = [
    {name = "cdr",email = "58159483+cdrg@users.noreply.github.com"}
]
license = {text = "LGPL-3.0-or-later"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "boto3 (>=1.37.24,<2.0.0)",
    "typed-ffmpeg (>=3.1.1,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "ttsmapi @ git+https://github.com/cdrg/ttsmapi.git",
    #"ttsmapi @ file:////home/cdr/projects/ttsmapi",
    "mypy-boto3-polly (>=1.40.13,<2.0.0)",
    "mypy-boto3-service-quotas (>=1.40.48,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
